Add a non-config starter for tomcat. All you need is the classpath including tomcat and the servlets - no config file. There is a requirement for a temp directory - the spec requires that webapps are provided with one I believe. It is possible to remove this or do it lazily or have some default in /tmp, if anyone wants to.
Also adding a context requires some base dir ( can be an empty one ) - so the default servlets points to something.
This can also be eliminated by setting a different default servlet - needs to be done if wanted.

It is also pretty fast - the unit tests show startup + first request in ~1 sec. 



git-svn-id: http://svn.apache.org/repos/asf/tomcat/trunk@677640 13f79535-47bb-0310-9956-ffa450edef68