Remove recycle of B2CConverter when used for URI conversion.
Recycle was an empty method until recently. It now eats up
all bytes in the ByteChunk underlying the B2CConverter in
order to fix BZ44494. That's fine and necessary for the
B2CConverter use in InputBuffer at the end of request handling,
but here, in the case of URI conversion, the associated
ByteChunk comes from a convert call of a previous request
and we will destroy a ByteChunk we still need to use outside of the
B2CConverter.


git-svn-id: http://svn.apache.org/repos/asf/tomcat/trunk@688857 13f79535-47bb-0310-9956-ffa450edef68