Provide documentation for deployer.
We do not have "printer" versions of our documentation pages anymore,
so I am using the common version.
Also, I removed "failonerror=false" flag, and removed renaming of the document. I see no point in renaming it. It was moved to the root folder, to remain relative to the images.

git-svn-id: http://svn.apache.org/repos/asf/tomcat/trunk@783863 13f79535-47bb-0310-9956-ffa450edef68