Fix EOLN issues in md5 files generated for our releases.
1. Do not add an EOLN after checksum value.
This was caused by difference in behaviour between Ant 1.6 and 1.7. Ant 1.6 did not add the newline.
We could use 'pattern' or 'format' attributes of the <checksum> task, but those are since Ant 1.7 only, not available in 1.6.5. This solution siplifies code and preserves compatibility with Ant 1.6.

2. Do not add an EOLN after the file name.
I removed ${line.separator} from the echoed text.
If the build is generated on Windows and verified on *nix, extra CR in a file name will cause complaints. I won't object to using an explicit LF here, though.

git-svn-id: http://svn.apache.org/repos/asf/tomcat/trunk@893375 13f79535-47bb-0310-9956-ffa450edef68