Add an HTTP header parser. The driver for this was an attempt to fix https://issues.apache.org/bugzilla/show_bug.cgi?id=52811
Parsing HTTP headers as per RFC2616 is not always as simple as it first appears. For headers that only use tokens the simple approach will normally be sufficient. However, for the other headers, while simple code meets 99.9% of cases, there are often some edge cases that make things far more complicated.

The purpose of this parser is to let the parser worry about the edge cases. It provides strict parsing of HTTP header values assuming that wrapped header lines have already been unwrapped. (The Tomcat header processing code does the unwrapping.)

The parser currently supports parsing of the following HTTP header values as per RFC 2616:
 - Content-Type

Support for additional headers will be provided as required. A quick scan of the Tomcat code base suggested a couple of places where using this parser may be useful such as  Ranges in the default servlet but there was not - at this point - a compelling case for immediate replacement. The expectation is that as problems are identified in header parsing, the fix will typically extend this parser to support the problematic header and then use the parser rather than custom code.

git-svn-id: http://svn.apache.org/repos/asf/tomcat/trunk@1300154 13f79535-47bb-0310-9956-ffa450edef68