For https://issues.apache.org/bugzilla/show_bug.cgi?id=53119 Port r1344253 to AjpNioProcessor
Prevent possible overflow exception with buffer in AjpNioProcessor.output() if it is called again after the previous write failed with IOException.

This is based on review of the issue fixed by r1344253. I see several reasons why it is hard to observe in Nio, but it is easy to prevent this issue. An overflow exception is bad because it is RuntimeException and is generally unexpected. This change:
- avoids writing to the buffer if att == null
- always clears the buffer. Note that Buffer.clear() method is rather lightweight one.

git-svn-id: http://svn.apache.org/repos/asf/tomcat/trunk@1346365 13f79535-47bb-0310-9956-ffa450edef68